# documentation/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Intra API Documentation",
    "description": "Welcome to the API Documentation for the Intra application. This API provides a set of endpoints to interact with various features of the application. Below are the different categories and their corresponding endpoints:\n"
  },
  "servers": [
    {
      "url": "https://le-lien-de-lapi"
    }
  ],
...
  "paths": {
...
    "/TEPITEK/{PARAMETERS}": {
      "get": {
        "summary": "Get Tepiteks results",
        "description": "Retrieves the Tepiteks results of a user based on the specified parameters.",
        "tags": [
          "English"
        ],
        "parameters": [
          {
            "name": "PARAMETERS",
            "in": "path",
            "description": "The parameters for Tepiteks results retrieval. Valid values are 'Year' or 'Last'.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Year",
                "Last"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/PLANNING/{START}/{END}": {
      "get": {
        "summary": "Get planning data",
        "description": "Retrieves the planning data within a specified time frame.",
        "tags": [
          "Planning"
        ],
        "parameters": [
          {
            "name": "START",
            "in": "path",
            "description": "The start date of the planning period.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "END",
            "in": "path",
            "description": "The end date of the planning period.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "API-Key",
        "description": "Provide your API key as a header parameter. You can obtain the API key by authenticating with the application.\n"
      }
    }
  },
  "security": [
    {
      "APIKeyAuth": []
    }
  ]
}